pyproject.toml
[tool.poetry]
name = "Weather app"
version = "0.1.0"
description = "Tinkoff Python Fintech 2022 Template"
authors = ["Vasiliy Kovalev"]

[tool.poetry.dependencies]
python = "^3.8"
typer = "^0.4.0"
requests = "^2.27.1"
bs4 = "^0.0.1"
python-decouple = "^3.6"


[tool.poetry.dev-dependencies]
pylint = "^2.12.1"
pytest = "^6.2.4"
pytest-deadfixtures = "*"
pytest-mock = "*"
mypy = "^0.910"
pycodestyle = "^2.7.0"
types-toml = "^0.10.0"
black = "^21.7b0"
isort = "^5.10.1"
requests = "^2.26.0"
pytest-cov = "*"
pydocstyle = "^6.1.1"
flake8 = "*"
flake8-builtins = "*"
flake8-comprehensions = "*"
flake8-eradicate = "*"
flake8-isort = "*"
flake8-logging-format = "*"
flake8-print = "*"
flake8-pytest = "*"
flake8-pytest-style = "*"
flake8-quotes = "*"
flake8-return = "*"
unify = "*"
safety = "*"
mock = "^4.0.3"
coverage = "^6.3.2"
types-requests = "^2.27.11"
autoflake = "^1.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
